+++ app/routes/__init__.py
@@ -17,7 +17,6 @@ from .namespaces.shift_report_material_ns import shift_report_material_ns
 from .namespaces.subscrtiption_ns import subscription_ns
 from .namespaces.template_ns import template_ns
 from .namespaces.user_ns import user_ns
-from .namespaces.work_category_ns import work_category_ns
 from .namespaces.work_material_relation_ns import work_material_relation_ns
 
 
@@ -32,7 +31,6 @@ def register_namespaces(api: Api):
     api.add_namespace(user_ns)
     api.add_namespace(city_ns)
     api.add_namespace(object_status_ns)
-    api.add_namespace(work_category_ns)
     api.add_namespace(object_ns)
     api.add_namespace(project_ns)
     api.add_namespace(material_ns)