+++ app/web/__init__.py
@@ -4,6 +4,7 @@ from .leaves.routes import leave_ns
 from .shift_reports.routes import shift_report_details_ns, shift_report_ns
 from .project_materials.routes import project_material_ns
 from .project_works.routes import project_work_ns
+from .work_categories.routes import work_category_ns
 from .work_prices.routes import work_price_ns
 from .works.routes import work_ns
 
@@ -14,5 +15,6 @@ def register_namespaces(api: Api):
     api.add_namespace(shift_report_details_ns)
     api.add_namespace(project_material_ns)
     api.add_namespace(project_work_ns)
+    api.add_namespace(work_category_ns)
     api.add_namespace(work_price_ns)
     api.add_namespace(work_ns)